requirements.txt
Django==2.0.3
django-haystack==2.8.0
elasticsearch==2.4.1
pytz==2018.3
urllib3==1.22
